refactor(frontend): enhance essay answers fetching with transformation utility
- Added a new utility function to transform infinite query data for posts with essay answers and likes.
- Updated the `usePostsEssayAnswersWithLikesInfinityQuery` hook to include a `select` option for data transformation.
- Refactored the AllAnswers component to utilize the new transformation utility, improving data handling and rendering logic.
- Adjusted the default page item count for better pagination control.
