+++ packages/frontend/src/modules/idea-hub/all-answers/index.tsx
@@ -4,16 +4,16 @@ import {
   PostEssayAnswerOrderByInput,
   PostOrderByInput,
 } from '__generated__/types'
-import { memo, useEffect, useMemo, useRef, useState } from 'react'
+import { memo, useEffect, useRef, useState } from 'react'
 import { toast } from 'sonner'
 
+import { DEFAULT_PAGE_ITEM_COUNT } from '@/api-utils/react-query/constants'
 import { usePostsEssayAnswersWithLikesInfinityQuery } from '@/api-utils/react-query/hooks/post'
 
-import { PostWithTwoTopLikesAnswersPerQuestion } from '../types'
 import PostAnswerCard from './post-answer-card'
 import PostAnswerCardSkeleton from './post-answer-card-skeleton'
+import { transformInfinitePostsEssayAnswersWithLikesDataToPosts } from './utils'
 
-const PAGE_SIZE = 5
 const ANSWER_TAKE = 2
 
 type AllAnswersProps = {
@@ -26,94 +26,28 @@ function AllAnswers({ onOpenModal }: AllAnswersProps) {
   const titleRef = useRef<HTMLDivElement>(null)
   const [hasShownToast, setHasShownToast] = useState(false)
 
-  const { data, isLoading, isFetchingNextPage, hasNextPage, fetchNextPage } =
-    usePostsEssayAnswersWithLikesInfinityQuery({
-      orderBy: [{ publishedDate: 'desc' }] as PostOrderByInput[],
-      take: PAGE_SIZE,
-      answerOrderBy: [{ likesCount: 'desc' }] as PostEssayAnswerOrderByInput[],
-      answerTake: ANSWER_TAKE,
-      where: {
-        postEssayQuestions: {
-          some: {
-            answers: {
-              some: {},
-            },
+  const {
+    data: posts,
+    isLoading,
+    isFetchingNextPage,
+    hasNextPage,
+    fetchNextPage,
+  } = usePostsEssayAnswersWithLikesInfinityQuery({
+    orderBy: [{ publishedDate: 'desc' }] as PostOrderByInput[],
+    take: DEFAULT_PAGE_ITEM_COUNT,
+    answerOrderBy: [{ likesCount: 'desc' }] as PostEssayAnswerOrderByInput[],
+    answerTake: ANSWER_TAKE,
+    where: {
+      postEssayQuestions: {
+        some: {
+          answers: {
+            some: {},
           },
         },
       },
-    })
-
-  // Transform API response to match component expected type
-  const posts = useMemo(() => {
-    if (!data?.pages) return []
-
-    const allPosts: PostWithTwoTopLikesAnswersPerQuestion['posts'] = []
-
-    data.pages.forEach((page) => {
-      if (!page) return
-
-      page.forEach((post) => {
-        if (!post) return
-
-        // Filter questions to only include those with answers
-        const questionsWithAnswers =
-          post.postEssayQuestions
-            ?.map((question) => ({
-              id: question.id,
-              title: question.title || '',
-              hint: question.hint || '',
-              answers:
-                question.answers?.map((answer) => {
-                  const member = answer?.member
-                  return {
-                    id: answer.id,
-                    content: answer.content || '',
-                    likesCount: answer.likesCount || 0,
-                    member: {
-                      id: member?.id || '',
-                      name: member?.name || '',
-                      nickname: member?.nickname || '',
-                      email: member?.email || '',
-                      avatar: {
-                        fileUrl: member?.avatar?.fileUrl || '',
-                        id: member?.avatar?.id || '',
-                      },
-                    },
-                  }
-                }) ?? [],
-            }))
-            .filter((question) => question.answers.length > 0) || []
-
-        // Only include posts that have at least one question with answers
-        if (questionsWithAnswers.length === 0) return
-        const postId = post.id || post.slug
-        if (!postId) {
-          throw new Error(
-            `Post is missing both id and slug. Data integrity issue detected. Post title: ${
-              post.title ?? 'N/A'
-            }.`
-          )
-        }
-        allPosts.push({
-          id: postId,
-          slug: post.slug || '',
-          title: post.title || '',
-          heroImage: {
-            resized: {
-              medium: post.heroImage?.resized?.medium || '',
-            },
-          },
-          subSubcategoriesOrdered:
-            post.subSubcategoriesOrdered?.map((cat) => ({
-              name: cat?.name || '',
-            })) || [],
-          postEssayQuestions: questionsWithAnswers,
-        })
-      })
-    })
-
-    return allPosts
-  }, [data])
+    },
+    select: transformInfinitePostsEssayAnswersWithLikesDataToPosts,
+  })
 
   // IntersectionObserver to trigger fetchNextPage
   useEffect(() => {
@@ -180,7 +114,7 @@ function AllAnswers({ onOpenModal }: AllAnswersProps) {
         ref={scrollContainerRef}
         className="flex scrollbar-thin snap-x snap-mandatory scroll-px-6 gap-6 overflow-x-auto px-6 tablet:scroll-px-8 tablet:px-8 desktop:scroll-px-12 desktop:px-12 hd:scroll-pr-14 hd:scroll-pl-[calc(50vw-600px+64px)] hd:pr-14 hd:pl-[calc(50vw-600px+64px)]"
       >
-        {posts.map((post) => (
+        {posts?.map((post) => (
           <div key={post.id} className="flex-shrink-0 snap-start">
             <PostAnswerCard post={post} onOpenModal={onOpenModal} />
           </div>
@@ -192,7 +126,7 @@ function AllAnswers({ onOpenModal }: AllAnswersProps) {
             aria-hidden="true"
           />
         )}
-        {!isLoading && posts.length === 0 && (
+        {!isLoading && posts?.length === 0 && (
           <div className="flex w-full items-center justify-center py-12 text-center">
             <p className="prose-p1 text-neutral-500">尚無回答</p>
           </div>
